pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=77.0"]

[project]
name = "adk_coder"
version = "0.0.1"
license = "Apache-2.0"
license-files = ["LICENSE"]
description = "Project short description."
readme = "README.md"
authors = [{ name = "Allen Porter", email = "allen.porter@gmail.com" }]
requires-python = ">=3.14"
classifiers = []
dependencies = [
    "python-dotenv>=1.0.0",
    "google-adk>=1.25.1",
    "click>=8.1.7",
    "pyyaml>=6.0",
    "rich>=13.7.0",
    "textual>=0.47.1",
    "portalocker>=2.8.2",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=6.0.0",
    "ty>=0.0.18",
    "ruff>=0.8.1",
    "pre-commit>=4.0.1",
    "coverage>=7.13.4",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
adk-coder = "adk_coder.main:main"

[project.urls]
Source = "https://github.com/allenporter/adk-coder"

[tool.setuptools.packages.find]
include = ["adk_coder*"]

[tool.setuptools.package-data]
"adk_coder.skills.builtin" = ["**/*"]



[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning:google.genai.types.*",
    "ignore::DeprecationWarning:google.adk.runners.*",
    "ignore::pytest.PytestUnhandledThreadExceptionWarning",
]
